+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/ide/inspections/highlight/ReassignedRobotVariableAnnotator.java
@@ -1,4 +1,4 @@
-package dev.xeonkryptos.xeonrobotframeworkplugin.ide.inspections.compilation;
+package dev.xeonkryptos.xeonrobotframeworkplugin.ide.inspections.highlight;
 
 import com.intellij.lang.annotation.AnnotationHolder;
 import com.intellij.lang.annotation.HighlightSeverity;
@@ -24,11 +24,11 @@ public class ReassignedRobotVariableAnnotator extends AbstractRobotVariableAnnot
             RobotVariableBodyId variableBodyId = RobotPsiImplUtil.getVariableBodyId(variable);
             if (variableBodyId != null && ((PsiPolyVariantReference) variableBodyId.getReference()).multiResolve(false).length > 0) {
                 ResolveResult[] resolveResults = ((PsiPolyVariantReference) variableBodyId.getReference()).multiResolve(false);
-                if (resolveResults.length > 1 && Arrays.stream(resolveResults)
-                                                       .filter(ResolveResult::isValidResult)
-                                                       .map(ResolveResult::getElement)
-                                                       .filter(result -> result instanceof RobotVariableDefinition)
-                                                       .count() > 1) {
+                if (Arrays.stream(resolveResults)
+                          .filter(ResolveResult::isValidResult)
+                          .map(ResolveResult::getElement)
+                          .filter(result -> result instanceof RobotVariableDefinition)
+                          .count() > 1) {
                     holder.newSilentAnnotation(HighlightSeverity.TEXT_ATTRIBUTES)
                           .textAttributes(DefaultLanguageHighlighterColors.REASSIGNED_LOCAL_VARIABLE)
                           .tooltip(RobotBundle.getMessage("annotation.variable.reassigned"))