+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/stub/index/KeywordCallNameIndex.java
@@ -6,6 +6,7 @@ import com.intellij.psi.stubs.StringStubIndexExtension;
 import com.intellij.psi.stubs.StubIndex;
 import com.intellij.psi.stubs.StubIndexKey;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.element.RobotKeywordCall;
+import dev.xeonkryptos.xeonrobotframeworkplugin.psi.stub.RobotStubFileElementType;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
@@ -30,5 +31,10 @@ public class KeywordCallNameIndex extends StringStubIndexExtension<RobotKeywordC
         StubIndexKey<String, RobotKeywordCall> stubIndexKey = getKey();
         String keywordNameInLowerCase = keywordName.toLowerCase();
         return StubIndex.getElements(stubIndexKey, keywordNameInLowerCase, project, scope, RobotKeywordCall.class);
+    }
+
+    @Override
+    public int getVersion() {
+        return RobotStubFileElementType.STUB_FILE_VERSION + super.getVersion() + 1;
     }
 }